refactor(CLI): Seclude schema of core commands
