+++ lib/classes/PluginManager.js
@@ -550,7 +550,7 @@ class PluginManager {
    * configuration file, throw an error
    */
   validateServerlessConfigDependency(command) {
-    if ('configDependent' in command && command.configDependent) {
+    if (command.configDependent || command.serviceDependencyMode === 'required') {
       if (!this.serverless.configurationInput) {
         const msg = [
           'This command can only be run in a Serverless service directory. ',
