+++ lib/plugins/aws/configCredentials.js
@@ -3,6 +3,7 @@
 const os = require('os');
 const credentials = require('./utils/credentials');
 const ServerlessError = require('../../serverless-error');
+const cliCommandsSchema = require('../../cli/commands-schema');
 
 class AwsConfigCredentials {
   constructor(serverless, options) {
@@ -16,27 +17,8 @@ class AwsConfigCredentials {
       config: {
         commands: {
           credentials: {
+            ...cliCommandsSchema.get('config credentials'),
             lifecycleEvents: ['config'],
-            options: {
-              key: {
-                usage: 'Access key for the provider',
-                shortcut: 'k',
-                required: true,
-              },
-              secret: {
-                usage: 'Secret key for the provider',
-                shortcut: 's',
-                required: true,
-              },
-              profile: {
-                usage: 'Name of the profile you wish to create. Defaults to "default"',
-                shortcut: 'n',
-              },
-              overwrite: {
-                usage: 'Overwrite the existing profile configuration in the credentials file',
-                shortcut: 'o',
-              },
-            },
           },
         },
       },
