+++ lib/plugins/aws/rollbackFunction.js
@@ -12,36 +12,6 @@ class AwsRollbackFunction {
 
     Object.assign(this, validate);
 
-    this.commands = {
-      rollback: {
-        commands: {
-          function: {
-            usage: 'Rollback the function to a specific version',
-            lifecycleEvents: ['rollback'],
-            options: {
-              'function': {
-                usage: 'Name of the function',
-                shortcut: 'f',
-                required: true,
-              },
-              'function-version': {
-                usage: 'Version of the function',
-                required: true,
-              },
-              'stage': {
-                usage: 'Stage of the function',
-                shortcut: 's',
-              },
-              'region': {
-                usage: 'Region of the function',
-                shortcut: 'r',
-              },
-            },
-          },
-        },
-      },
-    };
-
     this.hooks = {
       'rollback:function:rollback': async () =>
         BbPromise.bind(this)
