+++ lib/plugins/config.js
@@ -3,6 +3,7 @@
 const BbPromise = require('bluebird');
 const config = require('@serverless/utils/config');
 const ServerlessError = require('../serverless-error');
+const cliCommandsSchema = require('../cli/commands-schema');
 const isTabTabCompletionSupported = require('../utils/tabCompletion/isSupported');
 const muteConsoleLog = require('../utils/log/muteConsoleLog');
 const tabtabOptions = require('../utils/tabCompletion/tabtabOptions');
@@ -30,26 +31,12 @@ class Config {
 
     this.commands = {
       config: {
-        usage: 'Configure Serverless',
-        configDependent: false,
+        ...cliCommandsSchema.get('config'),
         lifecycleEvents: ['config'],
-        options: {
-          autoupdate: {
-            usage: 'Turn on auto update mechanism (turn off via "--no-autoupdate")',
-          },
-        },
         commands: {
           credentials: {
-            usage: 'Configures a new provider profile for the Serverless Framework',
-            lifecycleEvents: ['config'],
+            // Command defined in AWS context
             validProviders,
-            options: {
-              provider: {
-                usage: `Name of the provider. Supported providers: ${humanReadableProvidersList}`,
-                required: true,
-                shortcut: 'p',
-              },
-            },
           },
         },
       },
@@ -60,23 +47,11 @@ class Config {
         usage: 'Configure command line <tab> completion',
         commands: {
           install: {
-            usage: 'Install a <tab> completion for chosen shell',
+            ...cliCommandsSchema.get('config tabcompletion install'),
             lifecycleEvents: ['install'],
-            options: {
-              shell: {
-                usage:
-                  'Shell for which <tab> completion should be installed. ' +
-                  'Supported options: bash (default), zsh, fish ',
-                shortcut: 's',
-              },
-              location: {
-                usage: 'Custom location for shell config',
-                shortcut: 'l',
-              },
-            },
           },
           uninstall: {
-            usage: 'Uninstall a <tab> completion',
+            ...cliCommandsSchema.get('config tabcompletion uninstall'),
             lifecycleEvents: ['uninstall'],
           },
         },
