+++ lib/plugins/create/create.js
@@ -6,95 +6,14 @@ const fse = require('fs-extra');
 const untildify = require('untildify');
 
 const ServerlessError = require('../../serverless-error');
+const cliCommandsSchema = require('../../cli/commands-schema');
 const createFromTemplate = require('../../utils/createFromTemplate');
+const recommendedTemplatesList = require('../../templates/recommended-list');
+const humanReadableTemplatesList = require('../../templates/recommended-list/human-readable');
 const download = require('../../utils/downloadTemplateFromRepo');
 const renameService = require('../../utils/renameService').renameService;
 const copyDirContentsSync = require('../../utils/fs/copyDirContentsSync');
 const dirExistsSync = require('../../utils/fs/dirExistsSync');
-
-// class wide constants
-const validTemplates = [
-  'aws-clojure-gradle',
-  'aws-clojurescript-gradle',
-  'aws-nodejs',
-  'aws-nodejs-docker',
-  'aws-nodejs-typescript',
-  'aws-alexa-typescript',
-  'aws-nodejs-ecma-script',
-  'aws-python',
-  'aws-python3',
-  'aws-python-docker',
-  'aws-groovy-gradle',
-  'aws-java-maven',
-  'aws-java-gradle',
-  'aws-kotlin-jvm-maven',
-  'aws-kotlin-jvm-gradle',
-  'aws-kotlin-jvm-gradle-kts',
-  'aws-kotlin-nodejs-gradle',
-  'aws-scala-sbt',
-  'aws-csharp',
-  'aws-fsharp',
-  'aws-go',
-  'aws-go-dep',
-  'aws-go-mod',
-  'aws-ruby',
-  'aws-provided',
-  'tencent-go',
-  'tencent-nodejs',
-  'tencent-python',
-  'tencent-php',
-  'azure-csharp',
-  'azure-nodejs',
-  'azure-nodejs-typescript',
-  'azure-python',
-  'cloudflare-workers',
-  'cloudflare-workers-enterprise',
-  'cloudflare-workers-rust',
-  'fn-nodejs',
-  'fn-go',
-  'google-nodejs',
-  'google-python',
-  'google-go',
-  'kubeless-python',
-  'kubeless-nodejs',
-  'knative-docker',
-  'openwhisk-java-maven',
-  'openwhisk-nodejs',
-  'openwhisk-php',
-  'openwhisk-python',
-  'openwhisk-ruby',
-  'openwhisk-swift',
-  'spotinst-nodejs',
-  'spotinst-python',
-  'spotinst-ruby',
-  'spotinst-java8',
-  'twilio-nodejs',
-  'aliyun-nodejs',
-  'plugin',
-
-  // this template is used to streamline the onboarding process
-  // it uses the Node.js runtime and AWS provider
-  'hello-world',
-];
-
-const humanReadableTemplateList = (() => {
-  let lastGroupName = null;
-  let result = '';
-  let lineCount = 0;
-  const templateGroupRe = /^([a-z0-9]+)(-|$)/;
-  for (const templateName of validTemplates) {
-    const groupName = templateName.match(templateGroupRe)[1];
-    if (groupName !== lastGroupName || lineCount === 8) {
-      result += `\n${' '.repeat(45)}"${templateName}"`;
-      lastGroupName = groupName;
-      lineCount = 1;
-    } else {
-      result += `, "${templateName}"`;
-      ++lineCount;
-    }
-  }
-  return result;
-})();
 
 const handleServiceCreationError = (error) => {
   if (error.code !== 'EACCESS') throw error;
@@ -113,29 +32,8 @@ class Create {
 
     this.commands = {
       create: {
-        usage: 'Create new Serverless service',
+        ...cliCommandsSchema.get('create'),
         lifecycleEvents: ['create'],
-        options: {
-          'template': {
-            usage: `Template for the service. Available templates: ${humanReadableTemplateList}`,
-            shortcut: 't',
-          },
-          'template-url': {
-            usage: 'Template URL for the service. Supports: GitHub, BitBucket',
-            shortcut: 'u',
-          },
-          'template-path': {
-            usage: 'Template local path for the service.',
-          },
-          'path': {
-            usage: 'The path where the service should be created (e.g. --path my-service)',
-            shortcut: 'p',
-          },
-          'name': {
-            usage: 'Name for the service. Overwrites the default name of the created service.',
-            shortcut: 'n',
-          },
-        },
       },
     };
 
@@ -199,10 +97,10 @@ class Create {
   createFromTemplate() {
     const notPlugin = this.options.template !== 'plugin';
 
-    if (validTemplates.indexOf(this.options.template) === -1) {
+    if (!recommendedTemplatesList.includes(this.options.template)) {
       const errorMessage = [
         `Template "${this.options.template}" is not supported.`,
-        ` Supported templates are: ${humanReadableTemplateList}.`,
+        ` Supported templates are: ${humanReadableTemplatesList}.`,
       ].join('');
       throw new ServerlessError(errorMessage);
     }
