+++ lib/plugins/deploy.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const ServerlessError = require('../serverless-error');
+const cliCommandsSchema = require('../cli/commands-schema');
 const generateAnalyticsPayload = require('../utils/analytics/generatePayload');
 const { report: analyticsReport } = require('../utils/analytics');
 const processBackendNotificationRequest = require('../utils/processBackendNotificationRequest');
@@ -12,8 +13,7 @@ class Deploy {
 
     this.commands = {
       deploy: {
-        usage: 'Deploy a Serverless service',
-        configDependent: true,
+        ...cliCommandsSchema.get('deploy'),
         lifecycleEvents: [
           'deprecated#cleanup->package:cleanup',
           'deprecated#initialize->package:initialize',
@@ -24,71 +24,17 @@ class Deploy {
           'deploy',
           'finalize',
         ],
-        options: {
-          'conceal': {
-            usage: 'Hide secrets from the output (e.g. API Gateway key values)',
-          },
-          'stage': {
-            usage: 'Stage of the service',
-            shortcut: 's',
-          },
-          'region': {
-            usage: 'Region of the service',
-            shortcut: 'r',
-          },
-          'package': {
-            usage: 'Path of the deployment package',
-            shortcut: 'p',
-          },
-          'verbose': {
-            usage: 'Show all stack events during deployment',
-            shortcut: 'v',
-          },
-          'force': {
-            usage: 'Forces a deployment to take place',
-          },
-          'function': {
-            usage: "Function name. Deploys a single function (see 'deploy function')",
-            shortcut: 'f',
-          },
-          'aws-s3-accelerate': {
-            usage: 'Enables S3 Transfer Acceleration making uploading artifacts much faster.',
-          },
-        },
         commands: {
           function: {
-            usage: 'Deploy a single function from the service',
+            ...cliCommandsSchema.get('deploy function'),
             lifecycleEvents: ['initialize', 'packageFunction', 'deploy'],
-            options: {
-              'function': {
-                usage: 'Name of the function',
-                shortcut: 'f',
-                required: true,
-              },
-              'stage': {
-                usage: 'Stage of the function',
-                shortcut: 's',
-              },
-              'region': {
-                usage: 'Region of the function',
-                shortcut: 'r',
-              },
-              'force': {
-                usage: 'Forces a deployment to take place',
-              },
-              'update-config': {
-                usage:
-                  'Updates function configuration, e.g. Timeout or Memory Size without deploying code', // eslint-disable-line max-len
-                shortcut: 'u',
-              },
-            },
           },
           list: {
-            usage: 'List deployed version of your Serverless Service',
+            ...cliCommandsSchema.get('deploy list'),
             lifecycleEvents: ['log'],
             commands: {
               functions: {
-                usage: 'List all the deployed functions and their versions',
+                ...cliCommandsSchema.get('deploy list functions'),
                 lifecycleEvents: ['log'],
               },
             },
