+++ lib/plugins/info.js
@@ -1,28 +1,15 @@
 'use strict';
 
+const cliCommandsSchema = require('../cli/commands-schema');
+
 class Info {
   constructor(serverless) {
     this.serverless = serverless;
 
     this.commands = {
       info: {
-        usage: 'Display information about the service',
-        configDependent: true,
+        ...cliCommandsSchema.get('info'),
         lifecycleEvents: ['info'],
-        options: {
-          stage: {
-            usage: 'Stage of the service',
-            shortcut: 's',
-          },
-          region: {
-            usage: 'Region of the service',
-            shortcut: 'r',
-          },
-          verbose: {
-            usage: 'Display Stack output',
-            shortcut: 'v',
-          },
-        },
       },
     };
   }
