+++ lib/plugins/install.js
@@ -2,6 +2,7 @@
 
 const BbPromise = require('bluebird');
 
+const cliCommandsSchema = require('../cli/commands-schema');
 const download = require('../utils/downloadTemplateFromRepo');
 
 class Install {
@@ -11,19 +12,8 @@ class Install {
 
     this.commands = {
       install: {
-        usage: 'Install a Serverless service from GitHub or a plugin from the Serverless registry',
+        ...cliCommandsSchema.get('install'),
         lifecycleEvents: ['install'],
-        options: {
-          url: {
-            usage: 'URL of the Serverless service on GitHub',
-            required: true,
-            shortcut: 'u',
-          },
-          name: {
-            usage: 'Name for the service',
-            shortcut: 'n',
-          },
-        },
       },
     };
