+++ lib/plugins/metrics.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const cliCommandsSchema = require('../cli/commands-schema');
+
 class Metrics {
   constructor(serverless, options) {
     this.serverless = serverless;
@@ -7,29 +9,8 @@ class Metrics {
 
     this.commands = {
       metrics: {
-        usage: 'Show metrics for a specific function',
-        configDependent: true,
+        ...cliCommandsSchema.get('metrics'),
         lifecycleEvents: ['metrics'],
-        options: {
-          function: {
-            usage: 'The function name',
-            shortcut: 'f',
-          },
-          stage: {
-            usage: 'Stage of the service',
-            shortcut: 's',
-          },
-          region: {
-            usage: 'Region of the service',
-            shortcut: 'r',
-          },
-          startTime: {
-            usage: 'Start time for the metrics retrieval (e.g. 1970-01-01)',
-          },
-          endTime: {
-            usage: 'End time for the metrics retrieval (e.g. 1970-01-01)',
-          },
-        },
       },
     };
   }
