+++ lib/plugins/package/package.js
@@ -2,6 +2,7 @@
 
 const BbPromise = require('bluebird');
 const path = require('path');
+const cliCommandsSchema = require('../../cli/commands-schema');
 const zipService = require('./lib/zipService');
 const packageService = require('./lib/packageService');
 
@@ -19,8 +20,7 @@ class Package {
 
     this.commands = {
       package: {
-        usage: 'Packages a Serverless service',
-        configDependent: true,
+        ...cliCommandsSchema.get('package'),
         lifecycleEvents: [
           'cleanup',
           'initialize',
@@ -31,20 +31,6 @@ class Package {
           'compileEvents',
           'finalize',
         ],
-        options: {
-          stage: {
-            usage: 'Stage of the service',
-            shortcut: 's',
-          },
-          region: {
-            usage: 'Region of the service',
-            shortcut: 'r',
-          },
-          package: {
-            usage: 'Output path for the package',
-            shortcut: 'p',
-          },
-        },
         commands: {
           function: {
             type: 'entrypoint',
