+++ lib/plugins/plugin/install.js
@@ -5,6 +5,7 @@ const childProcess = BbPromise.promisifyAll(require('child_process'));
 const fse = require('fs-extra');
 const path = require('path');
 const _ = require('lodash');
+const cliCommandsSchema = require('../../cli/commands-schema');
 const yamlAstParser = require('../../utils/yamlAstParser');
 const fileExists = require('../../utils/fs/fileExists');
 const pluginUtils = require('./lib/utils');
@@ -21,15 +22,8 @@ class PluginInstall {
       plugin: {
         commands: {
           install: {
-            usage: 'Install and add a plugin to your service',
+            ...cliCommandsSchema.get('plugin install'),
             lifecycleEvents: ['install'],
-            options: {
-              name: {
-                usage: 'The plugin name',
-                required: true,
-                shortcut: 'n',
-              },
-            },
           },
         },
       },
