+++ lib/plugins/plugin/list.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const BbPromise = require('bluebird');
+const cliCommandsSchema = require('../../cli/commands-schema');
 const pluginUtils = require('./lib/utils');
 
 class PluginList {
@@ -14,7 +15,7 @@ class PluginList {
       plugin: {
         commands: {
           list: {
-            usage: 'Lists all available plugins',
+            ...cliCommandsSchema.get('plugin list'),
             lifecycleEvents: ['list'],
           },
         },
