+++ lib/plugins/plugin/search.js
@@ -2,6 +2,7 @@
 
 const BbPromise = require('bluebird');
 const chalk = require('chalk');
+const cliCommandsSchema = require('../../cli/commands-schema');
 const pluginUtils = require('./lib/utils');
 
 class PluginSearch {
@@ -15,15 +16,8 @@ class PluginSearch {
       plugin: {
         commands: {
           search: {
-            usage: 'Search for plugins',
+            ...cliCommandsSchema.get('plugin search'),
             lifecycleEvents: ['search'],
-            options: {
-              query: {
-                usage: 'Search query',
-                required: true,
-                shortcut: 'q',
-              },
-            },
           },
         },
       },
