+++ lib/plugins/plugin/uninstall.js
@@ -5,6 +5,7 @@ const childProcess = BbPromise.promisifyAll(require('child_process'));
 const fse = require('fs-extra');
 const path = require('path');
 const _ = require('lodash');
+const cliCommandsSchema = require('../../cli/commands-schema');
 const yamlAstParser = require('../../utils/yamlAstParser');
 const pluginUtils = require('./lib/utils');
 
@@ -19,15 +20,8 @@ class PluginUninstall {
       plugin: {
         commands: {
           uninstall: {
-            usage: 'Uninstall and remove a plugin from your service',
+            ...cliCommandsSchema.get('plugin uninstall'),
             lifecycleEvents: ['uninstall'],
-            options: {
-              name: {
-                usage: 'The plugin name',
-                required: true,
-                shortcut: 'n',
-              },
-            },
           },
         },
       },
