+++ lib/plugins/print.js
@@ -5,6 +5,7 @@ const _ = require('lodash');
 const jc = require('json-cycle');
 const yaml = require('js-yaml');
 const ServerlessError = require('../serverless-error');
+const cliCommandsSchema = require('../cli/commands-schema');
 
 class Print {
   constructor(serverless, options) {
@@ -14,20 +15,8 @@ class Print {
 
     this.commands = {
       print: {
-        usage: 'Print your compiled and resolved config file',
-        configDependent: true,
+        ...cliCommandsSchema.get('print'),
         lifecycleEvents: ['print'],
-        options: {
-          format: {
-            usage: 'Print configuration in given format ("yaml", "json", "text"). Default: yaml',
-          },
-          path: {
-            usage: 'Optional period-separated path to print a sub-value (eg: "provider.name")',
-          },
-          transform: {
-            usage: 'Optional transform-function to apply to the value ("keys")',
-          },
-        },
       },
     };
     this.hooks = {
