+++ lib/plugins/remove.js
@@ -1,28 +1,15 @@
 'use strict';
 
+const cliCommandsSchema = require('../cli/commands-schema');
+
 class Remove {
   constructor(serverless) {
     this.serverless = serverless;
 
     this.commands = {
       remove: {
-        usage: 'Remove Serverless service and all resources',
-        configDependent: true,
+        ...cliCommandsSchema.get('remove'),
         lifecycleEvents: ['remove'],
-        options: {
-          stage: {
-            usage: 'Stage of the service',
-            shortcut: 's',
-          },
-          region: {
-            usage: 'Region of the service',
-            shortcut: 'r',
-          },
-          verbose: {
-            usage: 'Show all stack events during deployment',
-            shortcut: 'v',
-          },
-        },
       },
     };
   }
