+++ lib/plugins/rollback.js
@@ -1,48 +1,19 @@
 'use strict';
 
+const cliCommandsSchema = require('../cli/commands-schema');
+
 class Rollback {
   constructor(serverless) {
     this.serverless = serverless;
 
     this.commands = {
       rollback: {
-        usage: 'Rollback the Serverless service to a specific deployment',
-        configDependent: true,
+        ...cliCommandsSchema.get('rollback'),
         lifecycleEvents: ['initialize', 'rollback'],
-        options: {
-          timestamp: {
-            usage: 'Timestamp of the deployment (list deployments with `serverless deploy list`)',
-            shortcut: 't',
-            required: false,
-          },
-          verbose: {
-            usage: 'Show all stack events during deployment',
-            shortcut: 'v',
-          },
-        },
         commands: {
           function: {
-            usage: 'Rollback the function to the previous version',
+            ...cliCommandsSchema.get('rollback function'),
             lifecycleEvents: ['rollback'],
-            options: {
-              'function': {
-                usage: 'Name of the function',
-                shortcut: 'f',
-                required: true,
-              },
-              'function-version': {
-                usage: 'Version of the function',
-                required: true,
-              },
-              'stage': {
-                usage: 'Stage of the function',
-                shortcut: 's',
-              },
-              'region': {
-                usage: 'Region of the function',
-                shortcut: 'r',
-              },
-            },
           },
         },
       },
