+++ lib/plugins/slstats.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const config = require('@serverless/utils/config');
+const cliCommandsSchema = require('../cli/commands-schema');
 
 class SlStats {
   constructor(serverless, options) {
@@ -9,19 +10,8 @@ class SlStats {
 
     this.commands = {
       slstats: {
-        usage: 'Enable or disable stats',
-        configDependent: true,
+        ...cliCommandsSchema.get('slstats'),
         lifecycleEvents: ['slstats'],
-        options: {
-          enable: {
-            usage: 'Enable stats ("--enable")',
-            shortcut: 'e',
-          },
-          disable: {
-            usage: 'Disable stats ("--disable")',
-            shortcut: 'd',
-          },
-        },
       },
     };
