+++ lib/plugins/standalone.js
@@ -12,6 +12,7 @@ const ServerlessError = require('../serverless-error');
 const isStandaloneExecutable =
   require('../utils/isStandaloneExecutable') && process.platform !== 'win32';
 const standaloneUtils = require('../utils/standalone');
+const cliCommandsSchema = require('../cli/commands-schema');
 
 const pipeline = promisify(stream.pipeline);
 
@@ -26,18 +27,13 @@ module.exports = class Standalone {
 
     this.commands = {
       upgrade: {
+        ...cliCommandsSchema.get('upgrade'),
         isHidden: !isStandaloneExecutable,
-        usage: 'Upgrade Serverless',
         lifecycleEvents: ['upgrade'],
-        options: {
-          major: {
-            usage: 'Enable upgrade to a new major release',
-          },
-        },
       },
       uninstall: {
+        ...cliCommandsSchema.get('uninstall'),
         isHidden: !isStandaloneExecutable,
-        usage: 'Uninstall Serverless',
         lifecycleEvents: ['uninstall'],
       },
     };
