+++ lib/templates/recommended-list/index.js
@@ -1,0 +1,65 @@
+'use strict';
+
+module.exports = [
+  'aws-clojure-gradle',
+  'aws-clojurescript-gradle',
+  'aws-nodejs',
+  'aws-nodejs-docker',
+  'aws-nodejs-typescript',
+  'aws-alexa-typescript',
+  'aws-nodejs-ecma-script',
+  'aws-python',
+  'aws-python3',
+  'aws-python-docker',
+  'aws-groovy-gradle',
+  'aws-java-maven',
+  'aws-java-gradle',
+  'aws-kotlin-jvm-maven',
+  'aws-kotlin-jvm-gradle',
+  'aws-kotlin-jvm-gradle-kts',
+  'aws-kotlin-nodejs-gradle',
+  'aws-scala-sbt',
+  'aws-csharp',
+  'aws-fsharp',
+  'aws-go',
+  'aws-go-dep',
+  'aws-go-mod',
+  'aws-ruby',
+  'aws-provided',
+  'tencent-go',
+  'tencent-nodejs',
+  'tencent-python',
+  'tencent-php',
+  'azure-csharp',
+  'azure-nodejs',
+  'azure-nodejs-typescript',
+  'azure-python',
+  'cloudflare-workers',
+  'cloudflare-workers-enterprise',
+  'cloudflare-workers-rust',
+  'fn-nodejs',
+  'fn-go',
+  'google-nodejs',
+  'google-python',
+  'google-go',
+  'kubeless-python',
+  'kubeless-nodejs',
+  'knative-docker',
+  'openwhisk-java-maven',
+  'openwhisk-nodejs',
+  'openwhisk-php',
+  'openwhisk-python',
+  'openwhisk-ruby',
+  'openwhisk-swift',
+  'spotinst-nodejs',
+  'spotinst-python',
+  'spotinst-ruby',
+  'spotinst-java8',
+  'twilio-nodejs',
+  'aliyun-nodejs',
+  'plugin',
+
+  // this template is used to streamline the onboarding process
+  // it uses the Node.js runtime and AWS provider
+  'hello-world',
+];
