+++ test/unit/lib/plugins/aws/rollbackFunction.test.js
@@ -66,8 +66,6 @@ describe('AwsRollbackFunction', () => {
 
     it('should have hooks', () => expect(awsRollbackFunction.hooks).to.be.not.empty);
 
-    it('should have commands', () => expect(awsRollbackFunction.commands).to.be.not.empty);
-
     it('should set the provider variable to an instance of AwsProvider', () =>
       expect(awsRollbackFunction.provider).to.be.instanceof(AwsProvider));
